Stock Price
This code uses a random forest regressor from the sklearn library to train a model on a dataset of stock prices. The input features are all of the columns except for the Date and Close columns, and the target variable is the Close column. The data is split into training and testing sets, and the model is trained on the training set. The model is then used to make predictions on the testing set, and the mean absolute error (MAE) is calculated as a measure of accuracy.